requirements.txt
openai>=1.16.0
python-dotenv>=1.0.0
moviepy>=1.0.3
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
aiofiles>=23.2.1
